# templates/openapi-spec.yaml
openapi: 3.0.3
info:
  title: My API Name
  description: |
    Replace with your API description. What problem does it solve?
    Who is it for? Keep it to 2–3 sentences.
  version: 1.0.0
  contact:
    name: brocketdesign
    url: https://rapidapi.com/brocketdesign

servers:
  - url: https://your-api.onrender.com
    description: Production

security:
  - RapidAPIKey: []

paths:
  /health:
    get:
      summary: Health check
      operationId: health
      responses:
        "200":
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /endpoint:
    get:
      summary: Main endpoint
      description: Replace this with your endpoint description.
      operationId: endpoint
      parameters:
        - name: q
          in: query
          required: true
          description: Search / input query
          schema:
            type: string
            example: "hello world"
        - name: limit
          in: query
          required: false
          description: Maximum number of results (1–100)
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden — invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  securitySchemes:
    RapidAPIKey:
      type: apiKey
      in: header
      name: X-RapidAPI-Key

  schemas:
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: object
        cached:
          type: boolean
          example: false

    Error:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: "The 'q' parameter is required"